requirements.txt
Authlib==1.2.1
beanie==1.23.0
fastapi==0.103.2
midi2audio==0.1.1
mido==1.2.10
motor==3.2.0
numpy==1.23.5
pydantic==2.4.2
python_jose==3.3.0
PyYAML==6.0
scipy==1.11.3
starlette
torch
uvicorn==0.23.2
itsdangerous